// src/types/customerGroup.ts
export interface CustomerGroup {
    id: string;
    name: string;
    description: string | null;
    createdAt: string;
  }
  
  export interface PersonInCharge {
    id: string;
    name: string;
  }
  
  export interface CustomerGroupDetails {
    id: string;
    name: string;
    description: string | null;
    createdAt: string;
  }
  